serverside done pretesting
