+++ server/models/index.js
@@ -1,0 +1,3 @@
+const Thought = require('./Thought');
+
+module.exports = { Thought };
